refactor: replace scattered tag handling functions with shiny new TagManager
- Consolidated all tag-related logic into a new `TagManager` class because yeah, it was everywhere.
- Nuked the old `tags.js` and buried `tags-service.js` alongside it. Good riddance.
- Updated everything to use `tagManager` instead of half-baked direct imports.
- Added an internal event for category tag additions because events are the glue of chaos.

Tech debt:
- Still relying on jQuery. Embrace the legacy.
- Sorting logic in `TagManager` feels like spaghetti but hey, it works for now.

Gotchas:
- If you were somehow relying on the glory days of direct tag function imports, tough luck—they’re gone.
